requirements.txt
playwright>=1.47.0
openai>=1.52.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3